arm64: dts: rockchip: add H96 Max+ TV box support
The device tree is based on DTS files for similar devices supported by
LibreELEC and the DTB extracted from the H96 Max+ Android 9 image
(RK3328_DC_H96_Max+_HS2734C_9_20191031.2215.img).

Unfortunately, devfreq is not available for the GPU (lima dirver) yet.
As a workaround, we simply do not configure pwm1 and the corresponding
regulator.

Signed-off-by: Ezra Buehler <ezra@easyb.ch>